refactor(skeleton): name web component file component.tsx instead of shadow.tsx
In the skeleton architecture there is no longer a split between a shadow
wrapper and an inner light-DOM web component (the kol-*-wc components are
gone), so the shadow.tsx filename is a misnomer. Standardise the skeleton
blueprint interactive components on component.tsx, matching the _skeleton
reference (web-components/skeleton/component.tsx):

- Rename button/shadow.tsx, button-link/shadow.tsx, link/shadow.tsx and
  link-button/shadow.tsx to component.tsx.
- Update the imports in component-list.ts and the co-located spec files.
- Update the ARC42 and click-event concept docs accordingly.
