+++ packages/components/src/components/button-link/test/snapshot.spec.tsx
@@ -2,7 +2,7 @@ import { KolButtonLinkTag } from '../../../core/component-names';
 import type { ButtonLinkProps } from '../../../schema';
 import { executeSnapshotTests } from '../../../utils/testing';
 
-import { KolButtonLink } from '../shadow';
+import { KolButtonLink } from '../component';
 
 executeSnapshotTests<ButtonLinkProps>(
 	KolButtonLinkTag,
