+++ packages/components/src/components/component-list.ts
@@ -5,8 +5,8 @@ import { KolAlert } from './alert/shadow';
 import { KolAvatar } from './avatar/component';
 import { KolBadge } from './badge/shadow';
 import { KolBreadcrumb } from './breadcrumb/shadow';
-import { KolButtonLink } from './button-link/shadow';
-import { KolButton } from './button/shadow';
+import { KolButtonLink } from './button-link/component';
+import { KolButton } from './button/component';
 import { KolCard } from './card/shadow';
 import { KolCombobox } from './combobox/shadow';
 import { KolDetails } from './details/shadow';
@@ -28,8 +28,8 @@ import { KolInputRadio } from './input-radio/shadow';
 import { KolInputRange } from './input-range/shadow';
 import { KolInputText } from './input-text/shadow';
 import { KolKolibri } from './kolibri/shadow';
-import { KolLinkButton } from './link-button/shadow';
-import { KolLink } from './link/shadow';
+import { KolLinkButton } from './link-button/component';
+import { KolLink } from './link/component';
 import { KolMeter } from './meter/component';
 import { KolModal } from './modal/shadow';
 import { KolNav } from './nav/shadow';
