+++ packages/components/src/components/link-button/test/snapshot.spec.tsx
@@ -2,7 +2,7 @@ import { KolLinkButtonTag } from '../../../core/component-names';
 import type { LinkButtonProps } from '../../../schema';
 import { executeSnapshotTests } from '../../../utils/testing';
 
-import { KolLinkButton } from '../shadow';
+import { KolLinkButton } from '../component';
 
 const baseObj = { _href: 'https://example.com', _icons: 'codicon codicon-home', _label: 'Label', _target: 'self' };
