+++ packages/components/src/components/link/snapshot.spec.tsx
@@ -2,7 +2,7 @@ import { KolLinkTag } from '../../core/component-names';
 import type { LinkProps } from '../../schema';
 import { executeSnapshotTests } from '../../utils/testing';
 
-import { KolLink } from './shadow';
+import { KolLink } from './component';
 
 const baseObj = { _href: 'https://example.com', _icons: 'codicon codicon-home', _label: 'Label', _target: 'self' };
